# internal/docs/swagger.yaml
definitions:
...
  api.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/api.UserResponse'
      workspace:
        $ref: '#/definitions/api.WorkspaceResponse'
    required:
    - token
    - user
    type: object
  api.AutoTagSuggestionResponse:
    properties:
      asset_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      tag_name:
        type: string
    required:
    - asset_id
    - created_at
    - id
    - tag_name
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  api.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  api.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  api.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - created_at
    - email
    - id
    - name
    type: object
...
  api.ValidationErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    required:
    - error
    - fields
    type: object
...
  api.WorkspaceResponse:
    properties:
      auto_tag_enabled:
        type: boolean
      auto_tag_mode:
        type: string
      created_at:
        type: string
      download_log_retention_days:
        type: integer
      duplicate_detection_mode:
        type: string
      event_log_retention_days:
        type: integer
      exif_keep:
        type: boolean
      exif_keep_gps:
        type: boolean
      icon_asset_id:
        type: string
      icon_version_id:
        type: string
      id:
        type: string
      locked_taxonomy:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      version_retention_count:
        type: integer
    required:
    - auto_tag_enabled
    - auto_tag_mode
    - created_at
    - download_log_retention_days
    - duplicate_detection_mode
    - event_log_retention_days
    - exif_keep
    - exif_keep_gps
    - id
    - locked_taxonomy
    - name
    - updated_at
    - version_retention_count
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/assets/{id}/auto-tag:
    post:
      description: Enqueues an auto_tag job for the asset, bypassing the workspace's
        auto_tag_enabled setting.
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Ineligible MIME type or no AI provider configured
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manually trigger AI tag suggestions for an asset
      tags:
      - Assets
  /api/v1/assets/{id}/auto-tag/suggestions:
    delete:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dismiss every pending AI tag suggestion for an asset
      tags:
      - Assets
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/api.AutoTagSuggestionResponse'
              type: array
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List pending AI tag suggestions for an asset
      tags:
      - Assets
  /api/v1/assets/{id}/auto-tag/suggestions/{sid}:
    delete:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Suggestion ID
        in: path
        name: sid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Suggestion not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dismiss a pending AI tag suggestion
      tags:
      - Assets
  /api/v1/assets/{id}/auto-tag/suggestions/{sid}/accept:
    post:
      description: Applies the suggested tag to the asset and deletes the suggestion.
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Suggestion ID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Suggestion not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept a pending AI tag suggestion
      tags:
      - Assets
  /api/v1/assets/{id}/auto-tag/suggestions/accept-all:
    post:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept every pending AI tag suggestion for an asset
      tags:
      - Assets
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates the user with email and password. Returns a JWT token
        in the response body and sets an <code>auth_token</code> httpOnly cookie (7-day
        expiry). The workspace embedded in the response is the user's first workspace
        — use <code>POST /api/v1/workspace/switch</code> to change it.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/api.ValidationErrorResponse'
      summary: Log in
      tags:
      - Auth
...
  /auth/refresh:
    post:
      description: Reissues a fresh JWT with a new 7-day expiry for the authenticated
        user and workspace. The new token is returned in the response body and also
        set as an updated <code>auth_token</code> cookie. Call this endpoint before
        the current token expires to maintain a session without re-login.
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh auth token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with name, email, and password. A default
        workspace is automatically created for the new user. The response includes
        a JWT auth token set as an <code>auth_token</code> httpOnly cookie and also
        returned in the response body.
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/api.ValidationErrorResponse'
      summary: Register a new user
      tags:
      - Auth
...
swagger: "2.0"